[GH-20] Deal with empty file in merge. (#21)
For some plugin, the output file is not created when there is no actual
write. We need to simulate the scenario and make sure the default
implementation of the merge function in `TmpShuffleFile` could deal with
it correctly. Zero should be returned to create a correct index file.

Add test case to simulate the situation when no file is available if the
partition length is zero.

+++ src/main/java/com/memverge/splash/TmpShuffleFile.java
@@ -55,18 +55,24 @@ public interface TmpShuffleFile extends ShuffleFile {
     final OutputStream out = makeOutputStream();
     log.info("merge {} files into {}.", srcFiles.size(), getPath());
     final List<Long> lengths = srcFiles.stream().map(file -> {
-      Long copied = null;
-      try (final InputStream in = file.makeInputStream()) {
-        copied = (long) IOUtils.copy(in, out);
-      } catch (IOException e) {
-        log.error("merge input from {} to {} failed.",
-            file.getPath(), getPath(), e);
-      } finally {
-        try {
-          file.delete();
-        } catch (IOException e) {
-          log.warn("delete {} failed.", file.getPath(), e);
+      Long copied = 0L;
+      try {
+        if (file.exists()) {
+          try (final InputStream in = file.makeInputStream()) {
+            copied = (long) IOUtils.copy(in, out);
+          } catch (IOException e) {
+            log.error("merge input from {} to {} failed.",
+                file.getPath(), getPath(), e);
+          } finally {
+            try {
+              file.delete();
+            } catch (IOException e) {
+              log.warn("delete {} failed.", file.getPath(), e);
+            }
+          }
         }
+      } catch (IOException e) {
+        log.error("failed to check existence of {}", file.getPath(), e);
       }
       return copied;
     }).collect(Collectors.toList());
